+++ gcs/CHANGES.md
@@ -54,9 +54,10 @@
     fs.gs.auth.service.account.enable
     ```
 
-1.  Add support for refresh token grant flow.
+1.  Add support for a new authentication type USER_CREDENTIALS.
     Retrieve a refresh token using the authorisation code grant flow and set up the following properties:
     ```properties
+    fs.gs.auth.refresh.type=USER_CREDENTIALS
     fs.gs.auth.refresh.token=YOUR_OAUTH2_REFRESH_TOKEN
     fs.gs.auth.client.id=YOUR_OAUTH2_CLIENT_ID
     fs.gs.auth.client.secret=YOUR_OAUTH2_CLIENT_SECRET